refactor(renderer): SpeakerRenderStage owns its construction and refresh
Step 2: `build_crossover`, `build_unified_table` and the topology refresh
move from `SpatialRenderer` (construction.rs) into `SpeakerRenderStage::new`
/ `refresh_for_topology`, taking `RendererControl` as a parameter. Verbatim
code motion; full suite incl. goldens passes unchanged.
