+++ omniphony-renderer/renderer/src/spatial_renderer/construction.rs
@@ -7,15 +7,14 @@
 //! These are `&mut self`/associated methods that do not hold the live or
 //! channel-state guards, so unlike `render_frame` they extract cleanly.
 
-use super::{BandRenderer, SpatialRenderer, evaluation_build_config};
-use crate::crossover::{FreqBand, LR4CrossoverBank, compute_bands};
+use super::{SpatialRenderer, evaluation_build_config};
 use crate::live_params::{
     CartesianEvaluationParams, EvaluationLiveParams, LiveEvaluationMode, LiveParams,
     PolarEvaluationParams, PreferredEvaluationMode, RampMode, RenderTopology, RendererControl,
 };
 use crate::render_backend::{
-    DegenerateVbapBackend, EffectiveEvaluationMode, GainModel, MultiBandTable, RenderRequest,
-    VbapBackend, build_prepared_render_engine,
+    DegenerateVbapBackend, EffectiveEvaluationMode, GainModel, RenderRequest, VbapBackend,
+    build_prepared_render_engine,
 };
 use crate::spatial_vbap::{DistanceModel, VbapPanner, VbapTableMode};
 use crate::speaker_layout::SpeakerLayout;
@@ -449,111 +448,6 @@ impl SpatialRenderer {
         }
     }
 
-    /// Build crossover band engines from a speaker layout.
-    ///
-    /// Returns `(render_bands, Some(filter_bank))` when the layout defines finite crossover
-    /// edges on at least one speaker (producing ≥ 2 bands), or `(single_band, None)` when
-    /// no crossover is needed. `render_bands` always has at least one entry.
-    #[allow(clippy::too_many_arguments)]
-    fn build_crossover(
-        control: &Arc<RendererControl>,
-        layout: &crate::speaker_layout::SpeakerLayout,
-        num_speakers: usize,
-        sample_rate: u32,
-        prev_bands: &[BandRenderer],
-    ) -> Result<(Vec<BandRenderer>, Option<LR4CrossoverBank>)> {
-        // For each new band, reuse the matching previous band (same speaker subset)
-        // so an evaluation-only refresh can keep its triangulated gain model.
-        let make_renderer = |b: &FreqBand| {
-            let prev = prev_bands
-                .iter()
-                .find(|p| p.speaker_indices == b.speaker_indices);
-            BandRenderer::from_band(b, layout, num_speakers, control, prev)
-        };
-
-        let bands = compute_bands(layout);
-        if bands.len() <= 1 {
-            let render_bands = bands
-                .iter()
-                .map(make_renderer)
-                .collect::<Result<Vec<_>>>()?;
-            return Ok((render_bands, None));
-        }
-
-        let cutoffs: Vec<f32> = bands
-            .windows(2)
-            .map(|w| w[0].high_hz)
-            .filter(|f| f.is_finite())
-            .collect();
-
-        let filter_bank = LR4CrossoverBank::new(&cutoffs, sample_rate);
-        let render_bands = bands
-            .iter()
-            .map(make_renderer)
-            .collect::<Result<Vec<_>>>()?;
-
-        log::info!(
-            "Crossover enabled: {} bands, cutoffs = {:?} Hz",
-            bands.len(),
-            cutoffs
-        );
-
-        Ok((render_bands, Some(filter_bank)))
-    }
-
-    /// Merge the per-band cartesian tables into a single multi-band table so a
-    /// lookup localises the cell once for all bands. Returns `None` (→ per-band
-    /// path) unless there are several bands all backed by a cartesian evaluator.
-    fn build_unified_table(
-        render_bands: &[BandRenderer],
-        num_speakers: usize,
-    ) -> Option<MultiBandTable> {
-        if render_bands.len() <= 1 {
-            return None;
-        }
-        // Every band shares the active evaluation mode, so they are all cartesian
-        // or all polar. Try cartesian first; if any band has no cartesian view,
-        // fall through to the polar path. A band without an engine (< 3 speakers)
-        // has no precomputed table → no unified table (per-band path).
-        let mut cartesian = Vec::with_capacity(render_bands.len());
-        let mut all_cartesian = true;
-        for band in render_bands {
-            let engine = band.engine()?;
-            match engine.cartesian_parts() {
-                Some(parts) => cartesian.push((parts, band.speaker_indices.as_slice())),
-                None => {
-                    all_cartesian = false;
-                    break;
-                }
-            }
-        }
-        if all_cartesian {
-            let table = MultiBandTable::build_cartesian(&cartesian, num_speakers);
-            if table.is_some() {
-                log::info!(
-                    "Crossover: unified cartesian table built for {} bands",
-                    render_bands.len()
-                );
-            }
-            return table;
-        }
-        drop(cartesian);
-
-        let mut polar = Vec::with_capacity(render_bands.len());
-        for band in render_bands {
-            let engine = band.engine()?;
-            polar.push((engine.polar_parts()?, band.speaker_indices.as_slice()));
-        }
-        let table = MultiBandTable::build_polar(&polar, num_speakers);
-        if table.is_some() {
-            log::info!(
-                "Crossover: unified polar table built for {} bands",
-                render_bands.len()
-            );
-        }
-        table
-    }
-
     /// Assemble the `SpatialRenderer` struct from fully resolved components.
     ///
     /// Called by both `new` and `from_vbap` after each constructor has built its
@@ -569,14 +463,13 @@ impl SpatialRenderer {
     ) -> Result<Self> {
         let active_topology = control.active_topology();
         let topology_identity = std::sync::Arc::as_ptr(&active_topology) as usize;
-        let (render_bands, crossover_filter_bank) = Self::build_crossover(
+        let speaker_stage = super::SpeakerRenderStage::new(
             &control,
             &active_topology.speaker_layout,
+            topology_identity,
             num_speakers,
             sample_rate,
-            &[],
         )?;
-        let unified_table = Self::build_unified_table(&render_bands, num_speakers);
 
         Ok(Self {
             num_speakers,
@@ -594,26 +487,7 @@ impl SpatialRenderer {
             loudness_gain: std::sync::atomic::AtomicU32::new(1.0_f32.to_bits()),
             auto_gain_triggered: std::sync::atomic::AtomicBool::new(false),
             control,
-            speaker_stage: super::SpeakerRenderStage {
-                num_speakers,
-                sample_rate,
-                render_bands,
-                render_bands_topology_identity: topology_identity,
-                unified_table,
-                crossover_filter_bank,
-                crossover_filter_states: Vec::new(),
-                crossover_band_scratch: std::array::from_fn(|_| Vec::new()),
-                band_gains_scratch: Vec::new(),
-                interp_end_scratch: Vec::new(),
-                speaker_gains_buf: vec![0.0f32; num_speakers],
-                bed_routing_gains_buf: vec![0.0f32; num_speakers],
-                delay_lines: {
-                    let max_delay = (0.1 * sample_rate as f32) as usize; // 100 ms
-                    (0..num_speakers)
-                        .map(|_| crate::delay_line::DelayLine::new(max_delay))
-                        .collect()
-                },
-            },
+            speaker_stage,
             object_params_buf: Vec::new(),
             speaker_params_buf: vec![
                 crate::live_params::SpeakerLiveParams::default();
@@ -627,37 +501,5 @@ impl SpatialRenderer {
             binaural_gain_buf: Vec::new(),
             binaural_direct_buf: Vec::new(),
         })
-    }
-
-    pub(super) fn refresh_crossover_for_topology(
-        &mut self,
-        topology_identity: usize,
-        active_layout: &crate::speaker_layout::SpeakerLayout,
-    ) -> Result<()> {
-        if self.speaker_stage.render_bands_topology_identity == topology_identity {
-            return Ok(());
-        }
-
-        // Pass the current bands so an evaluation-only recompute (unchanged geometry
-        // generation) reuses each band's triangulated gain model and rebuilds only
-        // the evaluation wrapper, instead of re-triangulating every band.
-        let (render_bands, crossover_filter_bank) = Self::build_crossover(
-            &self.control,
-            active_layout,
-            self.num_speakers,
-            self.sample_rate,
-            &self.speaker_stage.render_bands,
-        )?;
-        self.speaker_stage.unified_table =
-            Self::build_unified_table(&render_bands, self.num_speakers);
-        self.speaker_stage.render_bands = render_bands;
-        self.speaker_stage.crossover_filter_bank = crossover_filter_bank;
-        self.speaker_stage.crossover_filter_states.clear();
-        self.speaker_stage
-            .crossover_band_scratch
-            .iter_mut()
-            .for_each(Vec::clear);
-        self.speaker_stage.render_bands_topology_identity = topology_identity;
-        Ok(())
     }
 }
